Added iron rocks generated by WorleyNoise

+++ src/file/Chunk.h
@@ -35,7 +35,7 @@ struct chunk_t
 namespace file {
 
 void loadChunk(const void* p7rg, int x, int z, int rx, int rz, chunk_t** ppChunk);
-void saveChunk(const chunk_t* pChunk, int x, int z, int rx, int rz, uint8_t const ** ppBuffer, size_t* pcBuffer);
+void saveChunk(const chunk_t* pChunk, int x, int z, int rx, int rz, uint8_t const ** ppBuffer, uint32_t* pcBuffer);
 void decodeChunk(chunk_t* pChunk);
 void compareChunks(const chunk_t* pChunk0, const chunk_t* pChunk1);
 void encodeChunk(chunk_t** ppChunk, int x, int z, int rx, int rz);
